(fix) Correctly handle offset variable
Turns out that the page option in the google-images request
actually works how I wish it did. I was thrown off by the fact
that it handles an offset of 0 and an offset of 1 the same way,
so I assumed that it was done in batches of 10, my mistake.

It all works well now though, AND the code is a lot nicer
looking for it.